+++ sdk/core/azure-core/inc/azure/core.hpp
@@ -11,12 +11,14 @@
  */
 
 // azure/core
+#include "azure/core/base64.hpp"
 #include "azure/core/context.hpp"
 #include "azure/core/credentials.hpp"
 #include "azure/core/datetime.hpp"
 #include "azure/core/dll_import_export.hpp"
 #include "azure/core/etag.hpp"
 #include "azure/core/match_conditions.hpp"
+#include "azure/core/md5.hpp"
 #include "azure/core/nullable.hpp"
 #include "azure/core/request_conditions.hpp"
 #include "azure/core/response.hpp"