+++ sdk/core/azure-core/src/md5.cpp
@@ -1,0 +1,169 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// SPDX-License-Identifier: MIT
+
+#include "azure/core/md5.hpp"
+#include "azure/core/platform.hpp"
+
+#if defined(AZ_PLATFORM_WINDOWS)
+// Windows needs to go before bcrypt
+#include <windows.h>
+
+#include <bcrypt.h>
+#elif defined(AZ_PLATFORM_POSIX)
+#include <openssl/md5.h>
+#endif
+
+#include <stdexcept>
+#include <vector>
+
+namespace Azure { namespace Core {
+
+#if defined(AZ_PLATFORM_WINDOWS)
+
+  namespace Details {
+    struct AlgorithmProviderInstance
+    {
+      BCRYPT_ALG_HANDLE Handle;
+      std::size_t ContextSize;
+      std::size_t HashLength;
+
+      AlgorithmProviderInstance()
+      {
+        NTSTATUS status = BCryptOpenAlgorithmProvider(&Handle, BCRYPT_MD5_ALGORITHM, nullptr, 0);
+        if (!BCRYPT_SUCCESS(status))
+        {
+          throw std::runtime_error("BCryptOpenAlgorithmProvider failed");
+        }
+        DWORD objectLength = 0;
+        DWORD dataLength = 0;
+        status = BCryptGetProperty(
+            Handle,
+            BCRYPT_OBJECT_LENGTH,
+            reinterpret_cast<PBYTE>(&objectLength),
+            sizeof(objectLength),
+            &dataLength,
+            0);
+        if (!BCRYPT_SUCCESS(status))
+        {
+          throw std::runtime_error("BCryptGetProperty failed");
+        }
+        ContextSize = objectLength;
+        DWORD hashLength = 0;
+        status = BCryptGetProperty(
+            Handle,
+            BCRYPT_HASH_LENGTH,
+            reinterpret_cast<PBYTE>(&hashLength),
+            sizeof(hashLength),
+            &dataLength,
+            0);
+        if (!BCRYPT_SUCCESS(status))
+        {
+          throw std::runtime_error("BCryptGetProperty failed");
+        }
+        HashLength = hashLength;
+      }
+
+      ~AlgorithmProviderInstance() { BCryptCloseAlgorithmProvider(Handle, 0); }
+    };
+
+    struct Md5HashContext
+    {
+      std::string buffer;
+      BCRYPT_HASH_HANDLE hashHandle = nullptr;
+      std::size_t hashLength = 0;
+    };
+  } // namespace Details
+
+  Md5::Md5()
+  {
+    static Details::AlgorithmProviderInstance AlgorithmProvider{};
+
+    Details::Md5HashContext* md5Context = new Details::Md5HashContext;
+    m_md5Context = md5Context;
+    md5Context->buffer.resize(AlgorithmProvider.ContextSize);
+    md5Context->hashLength = AlgorithmProvider.HashLength;
+
+    NTSTATUS status = BCryptCreateHash(
+        AlgorithmProvider.Handle,
+        &md5Context->hashHandle,
+        reinterpret_cast<PUCHAR>(&md5Context->buffer[0]),
+        static_cast<ULONG>(md5Context->buffer.size()),
+        nullptr,
+        0,
+        0);
+    if (!BCRYPT_SUCCESS(status))
+    {
+      throw std::runtime_error("BCryptCreateHash failed");
+    }
+  }
+
+  Md5::~Md5()
+  {
+    Details::Md5HashContext* md5Context = static_cast<Details::Md5HashContext*>(m_md5Context);
+    BCryptDestroyHash(md5Context->hashHandle);
+    delete md5Context;
+  }
+
+  void Md5::Update(const uint8_t* data, std::size_t length)
+  {
+    Details::Md5HashContext* md5Context = static_cast<Details::Md5HashContext*>(m_md5Context);
+
+    NTSTATUS status = BCryptHashData(
+        md5Context->hashHandle,
+        reinterpret_cast<PBYTE>(const_cast<uint8_t*>(data)),
+        static_cast<ULONG>(length),
+        0);
+    if (!BCRYPT_SUCCESS(status))
+    {
+      throw std::runtime_error("BCryptHashData failed");
+    }
+  }
+
+  std::vector<uint8_t> Md5::Digest() const
+  {
+    Details::Md5HashContext* md5Context = static_cast<Details::Md5HashContext*>(m_md5Context);
+    std::vector<uint8_t> hash;
+    hash.resize(md5Context->hashLength);
+    NTSTATUS status = BCryptFinishHash(
+        md5Context->hashHandle,
+        reinterpret_cast<PUCHAR>(&hash[0]),
+        static_cast<ULONG>(hash.size()),
+        0);
+    if (!BCRYPT_SUCCESS(status))
+    {
+      throw std::runtime_error("BCryptFinishHash failed");
+    }
+    return hash;
+  }
+
+#elif defined(AZ_PLATFORM_POSIX)
+
+  Md5::Md5()
+  {
+    MD5_CTX* md5Context = new MD5_CTX;
+    m_md5Context = md5Context;
+    MD5_Init(md5Context);
+  }
+
+  Md5::~Md5()
+  {
+    MD5_CTX* md5Context = static_cast<MD5_CTX*>(m_md5Context);
+    delete md5Context;
+  }
+
+  void Md5::Update(const uint8_t* data, std::size_t length)
+  {
+    MD5_CTX* md5Context = static_cast<MD5_CTX*>(m_md5Context);
+    MD5_Update(md5Context, data, length);
+  }
+
+  std::vector<uint8_t> Md5::Digest() const
+  {
+    MD5_CTX* md5Context = static_cast<MD5_CTX*>(m_md5Context);
+    unsigned char hash[MD5_DIGEST_LENGTH];
+    MD5_Final(hash, md5Context);
+    return std::vector<uint8_t>(std::begin(hash), std::end(hash));
+  }
+
+#endif
+}} // namespace Azure::Core